requirements.txt
absl-py==2.1.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.2.0
babel==2.17.0
beartype==0.20.2
beautifulsoup4==4.13.3
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
cachetools==5.5.2
catalogue==2.0.10
catboost==1.2.7
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
dacite==1.9.2
databricks-sdk==0.53.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
docker==7.1.0
emoji==2.14.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
entrypoints==0.4
es_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.8.0/es_core_news_sm-3.8.0-py3-none-any.whl
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastai==2.8.0
fastapi==0.115.12
fastcore==1.8.0
fastdownload==0.0.7
fastjsonschema==2.21.1
fastprogress==1.0.3
fasttransform==0.0.1
filelock==3.18.0
Flask==3.1.0
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozendict==2.4.6
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.2.2
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
ImageHash==4.3.1
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.6.0
jaxlib==0.6.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.9.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
llvmlite==0.44.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.1.3
ml_dtypes==0.5.1
mlflow==2.22.0
mlflow-skinny==2.22.0
mlxtend==0.23.4
mpmath==1.3.0
multimethod==1.12
multitasking==0.0.11
murmurhash==1.0.12
namex==0.0.8
narwhals==1.30.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
opt_einsum==3.4.0
optree==0.14.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
peewee==3.17.9
pexpect==4.9.0
phik==0.12.4
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
plum-dispatch==2.5.7
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==4.25.7
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==4.0.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-json-logger==3.3.0
python-socketio==5.12.1
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==24.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sklearn-compat==0.1.3
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
sweetviz==2.3.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
thinc==8.3.4
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.2.0
typeguard==4.4.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.2
visions==0.8.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
xgboost==2.1.4
xlrd==2.0.1
yfinance==0.2.54
zipp==3.21.0
